Move config from root to src.
